rewrite query

+++ src/events/events.controller.ts
@@ -7,11 +7,13 @@ import {
   Patch,
   ValidationPipe,
   Logger,
+  NotFoundException,
+  ParseIntPipe,
 } from '@nestjs/common';
 import { Body, HttpCode, Param } from '@nestjs/common/decorators';
 import { CreateEventDto } from './create-event.dto';
 import { UpdateEventDto } from './update-event.dto';
-import { Like, MoreThan, Repository } from 'typeorm';
+import { FindOneOptions, Like, MoreThan, Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 
 @Controller('/events')
@@ -23,26 +25,6 @@ export class EventsController {
     private readonly repository: Repository<Event>,
   ) { }
 
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [
-        {
-          id: MoreThan(3), //{ id: 3 },
-          when: MoreThan(new Date('2020-12-12T13:00:00')),
-        },
-        {
-          description: Like('%meet%'),
-        },
-      ],
-      take: 2, //limit the number of result,
-      order: {
-        id: 'ASC',
-      },
-    });
-  }
-
   @Get()
   async findAll() {
     this.logger.log('Hit the findAll route');
@@ -51,9 +33,40 @@ export class EventsController {
     return events;
   }
 
+  // @Get('/practice')
+  // async practice() {
+  //   return await this.repository.find({
+  //     select: ['id', 'when'],
+  //     where: [
+  //       {
+  //         id: MoreThan(3), //{ id: 3 },
+  //         when: MoreThan(new Date('2020-12-12T13:00:00')),
+  //       },
+  //       {
+  //         description: Like('%meet%'),
+  //       },
+  //     ],
+  //     take: 2, //limit the number of result,
+  //     order: {
+  //       id: 'ASC',
+  //     },
+  //   });
+  // }
+
   @Get(':id')
-  async findOne(@Param('id') id) {
-    return await this.repository.findOne(id);
+  async findOne(@Param('id', ParseIntPipe) id: number) {
+    this.logger.log(`Hit the findOne route for id: ${id}`);
+
+    const options: FindOneOptions = {
+      where: {
+        id: id,
+      },
+    };
+    const event = await this.repository.findOne(options);
+
+    if (!event) throw new NotFoundException();
+
+    return event;
   }
 
   @Post()
@@ -68,6 +81,8 @@ export class EventsController {
   async update(@Param('id') id, @Body() input: UpdateEventDto) {
     const event = await this.repository.findOne(id);
 
+    if (!event) throw new NotFoundException();
+
     return await this.repository.save({
       ...event,
       ...input,
@@ -79,6 +94,9 @@ export class EventsController {
   @HttpCode(204)
   async remove(@Param('id') id) {
     const event = await this.repository.findOne(id);
+
+    if (!event) throw new NotFoundException();
+
     await this.repository.remove(event);
   }
 }
